\docType{package}
\name{etec16s-package}
\alias{etec16s-package}
\title{Dataset of 16S samples from 12 subjects challenged with enterotoxigenic Escherichia coli.}
\description{
  Data used in 'Individual-specific changes in the human gut microbiota after challenge with enterotoxigenic Escherichia coli and subsequent ciprofloxacin treatment'. Measurements are number of reads clustered into OTUs (operatonal taxanomic units) by DNAClust.
}
\author{
	Joseph N. Paulson
}
\references{
	Individual-specific changes in the human gut microbiota after challenge with enterotoxigenic Escherichia coli and subsequent ciprofloxacin treatment.
}

